添加按照id进行删除方法，增加校验。  -- 在分支上的提交

+++ src/main/java/com/hyc/service/GirlService.java
@@ -57,6 +57,8 @@ public class GirlService {
     }
 
     public void delete(Integer id) {
-        girlRepository.delete(id);
+        if (id != null) {
+            girlRepository.delete(id);
+        }
     }
 }
